Timmi/general fix

+++ include/struct.h
@@ -6,7 +6,7 @@
 /*   By: timmi <timmi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 10:41:49 by timmi             #+#    #+#             */
-/*   Updated: 2025/04/24 12:01:08 by timmi            ###   ########.fr       */
+/*   Updated: 2025/04/24 16:34:31 by timmi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ typedef struct s_list
 typedef struct s_shell
 {
 	char	**env;
-	int		cmd_count;
+	size_t	cmd_count;
 	char	*line;
 	t_ast	*root_node;
 	t_list	*head;
